Moved websocket connection to activity
